danger push

+++ resources/views/livewire/sellings/show-sellings.blade.php
@@ -29,6 +29,7 @@
                 <div
                     class="flex flex-col items-stretch justify-end flex-shrink-0 w-full space-y-2 md:w-auto md:flex-row md:space-y-0 md:items-center md:space-x-3">
                     <div class="flex items-stretch justify-end w-full space-x-3 md:w-auto">
+                        <button class="px-3 py-1 text-sm font-medium bg-white border border-gray-200 rounded-lg dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-100 dark:hover:text-white dark:hover:bg-gray-700 hover:cursor-pointer" wire:click="$dispatch('refreshSellings')">Refresh</button>
                         {{-- menambah w-full untuk satu component agar fit dengan card --}}
                         <div class="w-full dropdown sm:dropdown-end">
                             <label tabindex="0"
@@ -38,7 +39,7 @@
                                     <path clip-rule="evenodd" fill-rule="evenodd"
                                         d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" />
                                 </svg>
-                            {{ $form->showPerPage }}</label>
+                                {{ $form->showPerPage }}</label>
                             <ul tabindex="0"
                                 class="dropdown-content top-12 z-[1] menu p-2 shadow bg-base-100 rounded-box w-52">
                                 <li><a wire:click="setHowMuchPageShow(10)">10</a></li>
@@ -65,6 +66,7 @@
                             <th scope="col" class="px-4 py-3">Nama Produk</th>
                             <th scope="col" class="px-4 py-3">Jumlah</th>
                             <th scope="col" class="px-4 py-3">Satuan Pembelian</th>
+                            <th scope="col" class="px-4 py-3">Struk Pembayaran</th>
                             <th scope="col" class="px-4 py-3">
                                 <span class="sr-only">Actions</span>
                             </th>
@@ -94,6 +96,32 @@
                                     <td class="px-4 py-3">{{ $item->pivot->quantity }}</td>
                                     <td class="px-4 py-3">{{ $item->pivot->purchase_unit }}</td>
                                     @if ($i < 1)
+                                        <td class="text-center" rowspan="{{ $selling->products->count() }}">
+                                            @if (!is_null($selling->receipt))
+                                                <button onclick="my_modal_2.showModal()" wire:click="getDataForQr({{ $selling }})"
+                                                    class="p-2 bg-white border border-gray-200 rounded-lg shadow-sm dark:bg-gray-800 dark:border-gray-700">
+                                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none"
+                                                        viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor"
+                                                        class="w-5 h-5">
+                                                        <path stroke-linecap="round" stroke-linejoin="round"
+                                                            d="M3.75 4.875c0-.621.504-1.125 1.125-1.125h4.5c.621 0 1.125.504 1.125 1.125v4.5c0 .621-.504 1.125-1.125 1.125h-4.5A1.125 1.125 0 013.75 9.375v-4.5zM3.75 14.625c0-.621.504-1.125 1.125-1.125h4.5c.621 0 1.125.504 1.125 1.125v4.5c0 .621-.504 1.125-1.125 1.125h-4.5a1.125 1.125 0 01-1.125-1.125v-4.5zM13.5 4.875c0-.621.504-1.125 1.125-1.125h4.5c.621 0 1.125.504 1.125 1.125v4.5c0 .621-.504 1.125-1.125 1.125h-4.5A1.125 1.125 0 0113.5 9.375v-4.5z" />
+                                                        <path stroke-linecap="round" stroke-linejoin="round"
+                                                            d="M6.75 6.75h.75v.75h-.75v-.75zM6.75 16.5h.75v.75h-.75v-.75zM16.5 6.75h.75v.75h-.75v-.75zM13.5 13.5h.75v.75h-.75v-.75zM13.5 19.5h.75v.75h-.75v-.75zM19.5 13.5h.75v.75h-.75v-.75zM19.5 19.5h.75v.75h-.75v-.75zM16.5 16.5h.75v.75h-.75v-.75z" />
+                                                    </svg>
+
+                                                </button>
+                                            @else
+                                                <button type="button" wire:click="printReceipt({{ $selling }})"
+                                                    class="p-2 bg-white border border-gray-200 rounded-lg shadow-sm dark:bg-gray-800 dark:border-gray-700">
+                                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none"
+                                                        viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor"
+                                                        class="w-5 h-5">
+                                                        <path stroke-linecap="round" stroke-linejoin="round"
+                                                            d="M6.72 13.829c-.24.03-.48.062-.72.096m.72-.096a42.415 42.415 0 0110.56 0m-10.56 0L6.34 18m10.94-4.171c.24.03.48.062.72.096m-.72-.096L17.66 18m0 0l.229 2.523a1.125 1.125 0 01-1.12 1.227H7.231c-.662 0-1.18-.568-1.12-1.227L6.34 18m11.318 0h1.091A2.25 2.25 0 0021 15.75V9.456c0-1.081-.768-2.015-1.837-2.175a48.055 48.055 0 00-1.913-.247M6.34 18H5.25A2.25 2.25 0 013 15.75V9.456c0-1.081.768-2.015 1.837-2.175a48.041 48.041 0 011.913-.247m10.5 0a48.536 48.536 0 00-10.5 0m10.5 0V3.375c0-.621-.504-1.125-1.125-1.125h-8.25c-.621 0-1.125.504-1.125 1.125v3.659M18 10.5h.008v.008H18V10.5zm-3 0h.008v.008H15V10.5z" />
+                                                    </svg>
+                                                </button>
+                                            @endif
+                                        </td>
                                         <td rowspan="{{ $selling->products->count() }}" class="px-4 py-3 ">
                                             {{-- @can('menghapus catatan pengeluaran') --}}
                                             <div class="dropdown dropdown-left dropdown-end">
@@ -135,21 +163,35 @@
         </div>
     </div>
 
-    {{-- @can('menghapus catatan penjualan') --}}
     {{-- Modal Delete --}}
-    <dialog wire:ignore.self id="my_modal_1" class="modal">
+    @can('menghapus penjualan')
+        <dialog wire:ignore.self id="my_modal_1" class="modal">
+            <form method="dialog" class="modal-box">
+                <h3 class="text-lg font-bold">Menghapus Data Penjualan!</h3>
+                <p class="py-4">Apakah anda yakin ingin menghapus data
+                    <span class="font-medium text-red-600 text">{{ $form->transaction_code }}</span>?
+                </p>
+                <div class="modal-action">
+                    <button class="btn btn-error" wire:click="deleteSelling()">Hapus
+                    </button>
+                    <button class="btn btn-active">Tutup</button>
+                </div>
+            </form>
+        </dialog>
+    @endcan
+
+    {{-- Modal QR --}}
+    <dialog wire:ignore.self id="my_modal_2" class="modal">
         <form method="dialog" class="modal-box">
-            <h3 class="text-lg font-bold">Menghapus Data Penjualan!</h3>
-            <p class="py-4">Apakah anda yakin ingin menghapus data
-                <span class="font-medium text-red-600 text">{{ $form->transaction_code }}</span>?
-            </p>
+            <h3 class="mb-2 text-lg font-bold text-center">Silahkan Scan Barcode</h3>
+            <div class="flex items-center justify-center">
+                {{-- {!! QrCode::size(200)->generate(url('receipts/'.$this->form->receiptName)) !!} --}}
+                {!! QrCode::size(200)->generate(env('URL_NGROK').'/receipts/'.$this->form->receiptName) !!}
+            </div>
             <div class="modal-action">
-                <button class="btn btn-error" wire:click="deleteSelling()">Hapus
-                </button>
                 <button class="btn btn-active">Tutup</button>
             </div>
         </form>
     </dialog>
-    {{-- @endcan --}}
 
 </div>
